working on google authentication

+++ migrations/20210827180857-create-users.js
@@ -19,7 +19,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       local_account: {
-        type: Sequelize.SMALLINT 
+        type: Sequelize.INTEGER 
       },
       createdAt: {
         allowNull: false,
